refactor: split tree shake cjs tests
